Move create_render_context to render_util module
Move create_render_context and _build_rays from io.py into render_util.py. Move the corresponding unit tests from io_test.py into render_util_test.py, and re-export create_render_context from io.py for backward compatibility.
