+++ mujoco_warp/_src/io_jax_test.py
@@ -30,8 +30,8 @@ from absl.testing import parameterized
 
 import mujoco_warp as mjw
 from mujoco_warp import test_data
-from mujoco_warp._src.io import _is_array_spec
 from mujoco_warp._src.types import Callback
+from mujoco_warp._src.warp_util import is_array_spec
 
 _IO_TEST_MODELS = (
   "pendula.xml",
@@ -52,7 +52,7 @@ def _dims_match(test_obj, d1: Any, d2: Any, prefix: str = ""):
       _dims_match(test_obj, a1, a2, prefix + f1.name + ".")
       continue
 
-    if _is_array_spec(f1.type) or _is_array_spec(f2.type):
+    if is_array_spec(f1.type) or is_array_spec(f2.type):
       s1, s2 = a1.shape, a2.shape
       test_obj.assertEqual(len(s1), len(s2), f"{full_name} dims mismatch. Got {s1} and {s2}.")
       test_obj.assertEqual(s1, s2, f"{full_name} dims mismatch. Got {s1} and {s2}.")
@@ -91,7 +91,7 @@ def _check_type_matches_annotation(test_obj, obj: Any, prefix: str = ""):
       test_obj.assertIsInstance(np_scalar_type(val), type_, msg.format(**locals()))
       continue
 
-    if _is_array_spec(type_):
+    if is_array_spec(type_):
       test_obj.assertIsInstance(val, wp.array, msg.format(**locals()))
       continue
 
@@ -113,7 +113,7 @@ def _check_type_matches_annotation(test_obj, obj: Any, prefix: str = ""):
           test_obj.assertIsInstance(np_scalar_type(val), type_, msg.format(**locals()))
           continue
 
-        if _is_array_spec(type_):
+        if is_array_spec(type_):
           test_obj.assertIsInstance(val, wp.array, msg.format(**locals()))
           continue
 
@@ -138,7 +138,7 @@ def _check_annotation_compat(
     if v in (int, bool, float):
       continue
 
-    if _is_array_spec(v):
+    if is_array_spec(v):
       continue
 
     if wp.types.type_is_composite(v):
@@ -187,7 +187,7 @@ def _leading_dims_scale_w_nworld(test_obj, d1: Any, d2: Any, nworld1: int, nworl
       _leading_dims_scale_w_nworld(test_obj, a1, a2, nworld1, nworld2, prefix + f1.name + ".")
       continue
 
-    if _is_array_spec(f1.type) or _is_array_spec(f2.type):
+    if is_array_spec(f1.type) or is_array_spec(f2.type):
       s1, s2 = a1.shape[0], a2.shape[0]
       if s1 == s2:
         continue
